+++ internal/domain/dto/dtointerface.go
@@ -52,8 +52,8 @@ type Resource interface {
 }
 
 type Chunk interface {
-	GetLen() int
-	SetLen(len int)
+	GetLen() int64
+	SetLen(len int64)
 	GetData() []byte
 	SetData(data []byte)
 	GetError() error
